+++ docs/CHANGELOG.md
@@ -3,6 +3,15 @@
 ## 15.4.6
 
 ### 15.4.6 Features
+## 16.6.0
+
+### 16.6.0 Fixes
+
+- `[Datagrid]` Updated data type of cell in drilldown click event. ([EP#7473](https://github.com/infor-design/enterprise/issues/7473))
+
+## 16.5.0
+
+### 16.5.0 Fixes
 
 - `[General]` Fixed type issue
 ## 15.4.5